feat: #55 restric pages
