+++ src/components/common/navbar/organisms/AvatarDropDownMenu.tsx
@@ -1,3 +1,4 @@
+import { useCheckLoginStatus } from '@/components/login/CheckLoginStatus';
 import {
 	pageDashboardAdminRoute,
 	pageDashboardFavouriteRoute,
@@ -7,34 +8,13 @@ import {
 	pageLoginRoute,
 } from '@/pages';
 import Link from 'next/link';
-import { useEffect, useState } from 'react';
 
 type AvatarDropDownMenuProps = {
 	topPosition: string;
 };
 
 const AvatarDropDownMenu: React.FC<AvatarDropDownMenuProps> = ({ topPosition }) => {
-	const [isLoggedIn, setIsLoggedIn] = useState<boolean | undefined>(undefined);
-
-	useEffect(() => {
-		const checkLoginStatus = async () => {
-			try {
-				const response = await fetch('http://localhost:8181/api/users/me');
-				console.log('chujj ' + response.status);
-				if (response.status === 200) {
-					setIsLoggedIn(true);
-				} else if (response.status === 401) {
-					setIsLoggedIn(false);
-				} else {
-					console.error('Błąd podczas sprawdzania statusu logowania:', response.status);
-				}
-			} catch (error) {
-				console.error('Błąd podczas sprawdzania statusu logowania:', error);
-			}
-		};
-		checkLoginStatus();
-	}, []);
-	console.log('chujjjjj ' + isLoggedIn);
+	const isLoggedIn = useCheckLoginStatus();
 	return (
 		<div
 			className={`${topPosition} + bg-white2darkgrey h-[auto] w-[160px] absolute top-[80px] right-[0px] shadow-md rounded-lg p-2`}
