+++ src/components/login/CheckLoginStatus.tsx
@@ -1,0 +1,28 @@
+import { useState, useEffect } from 'react';
+
+export const useCheckLoginStatus = () => {
+	const [isLoggedIn, setIsLoggedIn] = useState<boolean | undefined>(undefined);
+
+	useEffect(() => {
+		const checkLoginStatus = async () => {
+			try {
+				const response = await fetch('http://localhost:8181/api/users/me');
+				console.log('response ' + response.status);
+				if (response.status === 200) {
+					setIsLoggedIn(true);
+				} else if (response.status === 401) {
+					setIsLoggedIn(false);
+				} else {
+					console.error('Błąd podczas sprawdzania statusu logowania:', response.status);
+				}
+			} catch (error) {
+				console.error('Błąd podczas sprawdzania statusu logowania:', error);
+			}
+		};
+		checkLoginStatus();
+	}, []);
+
+	console.log('isLoggedIn ' + isLoggedIn);
+
+	return isLoggedIn;
+};
